rm legacy_setting view, fix filter matgin, switchcompat color #BG-47
