--- app/src/main/res/layout/legacy_activity_setting.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:orientation="vertical"
-        android:layout_height="match_parent">
-
-
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:orientation="vertical"
-                android:layout_gravity="center_vertical"
-                android:layout_height="match_parent">
-
-                <include layout="@layout/toolbar"/>
-
-                <android.support.v7.widget.RecyclerView
-                    android:layout_width="match_parent"
-                    android:background="#f1f1f1"
-                    android:id="@+id/recycler_settinglist"
-                    android:layout_height="match_parent">
-
-                </android.support.v7.widget.RecyclerView>
-
-            </LinearLayout>
-
-            <include layout="@layout/progress_loading"/>
-        </FrameLayout>
-
-
-    </LinearLayout>
-
-</LinearLayout>
